InputHandler // Refactor handleInput() to triageInput().
